+++ src/Encamina.Enmarcha.SemanticKernel.Abstractions/IMemoryManager.cs
@@ -1,5 +1,7 @@
 ﻿// Ignore Spelling: Upsert
 
+using Encamina.Enmarcha.SemanticKernel.Abstractions.Events;
+
 using Microsoft.SemanticKernel;
 using Microsoft.SemanticKernel.Memory;
 
@@ -11,6 +13,12 @@ namespace Encamina.Enmarcha.SemanticKernel.Abstractions;
 public interface IMemoryManager
 {
     /// <summary>
+    /// This event is fired when <see cref="IMemoryManager"/> executes an action.
+    /// The event information is defined in the object <see cref="MemoryManagerEventArgs"/>, and the action types in the object <see cref="MemoryManagerEventTypes"/>.
+    /// </summary>
+    event EventHandler<MemoryManagerEventArgs> MemoryManagerEvent;
+
+    /// <summary>
     /// Gets the instance of the memory store manage by this manager.
     /// </summary>
     IMemoryStore MemoryStore { get; }
@@ -21,10 +29,17 @@ public interface IMemoryManager
     /// <param name="memoryId">The memory unique identifier.</param>
     /// <param name="collectionName">Name of the collection where the content will be saved.</param>
     /// <param name="chunks">List of strings containing the content of the memory.</param>
-    /// <param name="cancellationToken">Cancellation token to cancel the operation.</param>
+    /// <param name="kernel"><see cref="Kernel"/> instance object.</param>
     /// <param name="metadata">Metadata of the memory.</param>
+    /// <param name="cancellationToken">Cancellation token to cancel the operation.</param>
     /// <returns>A <see cref="Task"/> representing the asynchronous operation.</returns>
-    Task UpsertMemoryAsync(string memoryId, string collectionName, IEnumerable<string> chunks, Kernel kernel, IDictionary<string, string> metadata = null, CancellationToken cancellationToken = default);
+    Task UpsertMemoryAsync(
+        string memoryId,
+        string collectionName,
+        IEnumerable<string> chunks,
+        Kernel kernel,
+        IDictionary<string, string> metadata = null,
+        CancellationToken cancellationToken = default);
 
     /// <summary>
     /// Deletes the memory content from a collection.
@@ -53,6 +68,7 @@ public interface IMemoryManager
     /// The <c>key</c> in the dictionary must contain a unique identifier for the content of the memory,
     /// and the <c>value</c> of the dictionary must provide the memory content (chunks and metadata).
     /// </param>
+    /// <param name="kernel"><see cref="Kernel"/> object instance.</param>
     /// <param name="cancellationToken">Cancellation token to cancel the operation.</param>
     /// <returns>The unique identifiers for the memory records (not necessarily the same as the unique identifier of the memory content).</returns>
     IAsyncEnumerable<string> BatchUpsertMemoriesAsync(string collectionName, IDictionary<string, MemoryContent> memoryContents, Kernel kernel, CancellationToken cancellationToken);